Corrige erros de compilação do target Fazendinha
- Adiciona SCNNode.isDescendant(of:) (não existe nativo, é método de UIView/NSView)
- Expõe gameState (era private) para o FarmGameView acessar
- Resolve ambiguidade de .pi e mismatches CGFloat/Float em animações
  (ManuCharacter, DalmatianCharacter, NenaoCharacter, FarmWorldBuilder)

Encontrado ao rodar a run #1 do workflow "Fazendinha Build" no CI.

+++ Fazendinha/Characters/DalmatianCharacter.swift
@@ -75,5 +75,5 @@ final class DalmatianCharacter: SCNNode {
     }
 
     required init?(coder: NSCoder) { nil }
-    func wagTail(progress: CGFloat) { tailNode.eulerAngles.z = 0.4 + sin(progress * .pi * 4) * 0.3 }
+    func wagTail(progress: CGFloat) { tailNode.eulerAngles.z = Float(0.4 + sin(progress * .pi * 4) * 0.3) }
 }
